refactor and feat: split measure_IA into box and lightcone files and create manager functions for lightcone versions
