+++ src/measureia/__init__.py
@@ -1,8 +1,8 @@
 # import internal classes for use, so the module file names do not need to be called.
 
 # import base and wrapper classes
-from .measure_IA import MeasureIABox
-from .measure_IA import MeasureIALightcone
+from .measure_IA_box import MeasureIABox
+from .measure_IA_lightcone import MeasureIALightcone
 from .measure_IA_base import MeasureIABase
 
 # import covariance measurement class
